Rewrite in TypeScript
Source, tests, e2e specs, demo and config files are all TypeScript under
`strict` plus `noUncheckedIndexedAccess`. No runtime behaviour changed: all 23
unit tests and 7 browser tests pass unmodified.

Declarations are now generated from the source by `tsconfig.build.json` and
copied into both `lib/` and `es/`, replacing the hand-written `src/index.d.ts`.
Declaration maps ship as well, so "go to definition" resolves to the real
`.tsx` in the published `src/`. The public prop surface is unchanged from
1.0.0; `TextFieldProps` and `DateLikeWithMessage` are newly exported.

Two things the type checker surfaced that were wrong before:

- react-datepicker's props are a three-way union discriminated by
  `selectsRange` / `selectsMultiple`, with `onChange` receiving a Date, a tuple
  or an array depending on the variant. All three discriminants are optional,
  so an object without them matches every variant. The single-date member is
  now selected with `Extract` and pinned at the JSX site, making range and
  multi-select compile errors instead of a silent change to `onChange`.
- `showMonthYearDropdown` and `withPortal` are typed `true | never`, so "off"
  is the absence of the prop. Passing `false` was a type error; both are now
  omitted. They already defaulted to off, so behaviour is unchanged.

Dropped vitest-axe: it augments the legacy `Vi` global namespace that Vitest 4
no longer uses, so `toHaveNoViolations` was never actually typed. The unit
tests call axe-core directly and assert on violation ids, which also names the
offending rule on failure.

The packaging check earned its keep during this change - it caught that the
`files` negation patterns still said `.test.jsx` after the rename, which would
have shipped the test suite to consumers.
