+++ .github/workflows/ci.yml
@@ -30,7 +30,7 @@ jobs:
       - name: Lint
         run: npm run lint
 
-      - name: Typecheck published declarations
+      - name: Typecheck
         run: npm run typecheck
 
       - name: Test
