+++ AGENTS.md
@@ -3,8 +3,8 @@
 > Guidance for coding agents working in this repo. `CLAUDE.md` imports this
 > file, so edit **this** file — never `CLAUDE.md`.
 
-A date picker with Autodesk HIG styling, wrapping `react-datepicker`. Published
-to npm as `react-hig-datepicker`.
+A date picker with Autodesk HIG styling, wrapping `react-datepicker`. Written in
+TypeScript, published to npm as `react-hig-datepicker`.
 
 The repo went from unusable to 1.0 in one pass (see `CHANGELOG.md`). Most of
 what follows is there to stop that work being undone by a plausible-looking
@@ -17,7 +17,7 @@ npm start          # demo at localhost:5173, imports straight from src/
 npm test           # 23 unit tests (Vitest + Testing Library, jsdom)
 npm run test:e2e   # 7 browser tests (Playwright + axe); needs `npx playwright install chromium`
 npm run lint       # ESLint incl. jsx-a11y
-npm run typecheck  # compiles src/index.d.ts against types-test.ts
+npm run typecheck  # type-checks src, tests, e2e, demo and configs
 npm run build      # -> es/ (ESM), lib/ (CJS), css/datePicker.css
 npm run audit      # runtime deps only; must be 0 before any release
 npm run check:package  # asserts the tarball npm would publish is usable
@@ -85,6 +85,22 @@ map, so `react/jsx-runtime` is unresolvable there under Node ESM.
 reads as an argument separator and the font stack silently loses its fallback,
 rendering the calendar in the browser's default serif.
 
+**react-datepicker's props are a three-way union.** They are discriminated by
+`selectsRange` / `selectsMultiple`, and `onChange` receives a `Date`, a tuple or
+an array depending on the variant. All three discriminants are optional, so an
+object without them matches every variant and TypeScript reports against the
+last one. `SingleDateProps` picks the single-date member with `Extract`, and the
+JSX site passes both discriminants explicitly as `false` to pin it. Removing
+either is why `onChange` will suddenly appear to take `Date[]`.
+
+**Some react-datepicker props are typed `true | never`.**
+`showMonthYearDropdown` and `withPortal` model "off" as the *absence* of the
+prop, so passing `false` is a type error. Omit them.
+
+**Declarations are generated, never hand-written.** `tsconfig.build.json` emits
+them into `types/`, and the build copies that tree into both `lib/` and `es/`.
+Do not add a checked-in `index.d.ts` - that is what drifted before.
+
 **react-datepicker's CSS often outranks yours.** It uses selectors like
 `.react-datepicker__day:not([aria-disabled="true"]):hover` (specificity 0,3,0),
 which beats a plain `.react-datepicker__day:hover`. Match the specificity rather
@@ -120,6 +136,9 @@ in real Chromium. A green unit run is not evidence that contrast is fine.
   radius and the keyboard-cursor colour; none of it failed a unit test.
 - Prefer removing a dependency over updating it. A consumer install is currently
   13 packages; keep it that way.
+- Check that test tooling actually type-checks its own matchers. `vitest-axe`
+  silently stopped augmenting Vitest's types at v3 and its assertions became
+  `any`; the unit tests now call `axe-core` directly instead.
 - Verify visual claims with a real browser before saying the UX is unchanged.
   Four regressions in this repo's history were invisible to the test suite.
