+++ demo/index.html
@@ -7,6 +7,6 @@
   </head>
   <body>
     <div id="demo"></div>
-    <script type="module" src="./src/main.jsx"></script>
+    <script type="module" src="./src/main.tsx"></script>
   </body>
 </html>
