+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-hig-datepicker",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "An accessible date picker React component styled with the Autodesk HIG design language",
   "keywords": [
     "react",
@@ -46,8 +46,7 @@
     "es",
     "lib",
     "src",
-    "!src/**/*.test.js",
-    "!src/**/*.test.jsx",
+    "!src/**/*.test.*",
     "CHANGELOG.md"
   ],
   "engines": {
@@ -56,8 +55,8 @@
   "scripts": {
     "build": "node scripts/build.mjs",
     "clean": "node -e \"for (const d of ['lib','es','css','demo/dist']) require('node:fs').rmSync(d,{recursive:true,force:true})\"",
-    "start": "vite --config vite.demo.config.js",
-    "demo:build": "vite build --config vite.demo.config.js",
+    "start": "vite --config vite.demo.config.ts",
+    "demo:build": "vite build --config vite.demo.config.ts",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "test": "vitest run",
@@ -85,10 +84,12 @@
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.1.0",
     "@testing-library/user-event": "^14.5.2",
+    "@types/node": "^26.2.0",
     "@types/react": "^19.0.2",
     "@types/react-dom": "^19.0.2",
     "@vitejs/plugin-react": "^4.3.4",
     "@vitest/coverage-v8": "^4.1.10",
+    "axe-core": "^4.13.0",
     "eslint": "^9.17.0",
     "eslint-plugin-jsx-a11y": "^6.10.2",
     "eslint-plugin-react": "^7.37.3",
@@ -99,8 +100,8 @@
     "react-dom": "^19.0.0",
     "sass": "^1.83.0",
     "typescript": "^5.7.2",
+    "typescript-eslint": "^8.66.0",
     "vite": "^6.0.5",
-    "vitest": "^4.1.10",
-    "vitest-axe": "^0.1.0"
+    "vitest": "^4.1.10"
   }
 }
