--- src/TextField.jsx
@@ -1,183 +1,0 @@
-import { forwardRef, useCallback, useId, useState } from "react";
-import clsx from "clsx";
-import { ClearIcon } from "./icons.jsx";
-
-/**
- * A vendored, accessibility-corrected reimplementation of
- * `TextFieldPresenter` from @hig/text-field.
- *
- * The upstream package was deprecated with a `react@^16.3.2` peer range, and
- * its replacement (@hig/input) caps at react@^17 on top of an end-of-life
- * emotion 10. Either one makes this package uninstallable for React 18/19
- * consumers, so the markup and class names are reproduced here instead. The
- * DOM structure and `hig__text-field-v1*` class names are kept identical so
- * existing consumer style overrides continue to apply.
- *
- * react-datepicker clones this element and injects `ref`, `value`, `id`,
- * `readOnly`, `disabled`, `placeholder` and the event handlers, so the ref is
- * forwarded all the way to the `<input>` that react-datepicker focuses.
- */
-const TextField = forwardRef(function TextField(
-  {
-    clearButtonTitle = "Clear date",
-    disabled = false,
-    errors,
-    focused,
-    hideInstructionsOnErrors = false,
-    icon,
-    id,
-    instructions,
-    label,
-    onBlur,
-    onClearButtonClick,
-    onFocus,
-    // react-datepicker injects its own `readOnly`; the field is always a
-    // calendar trigger, so the injected value is ignored on purpose.
-    readOnly: _ignoredReadOnly,
-    required,
-    showClearButton = false,
-    type = "text",
-    value,
-    "aria-describedby": ariaDescribedBy,
-    ...inputProps
-  },
-  ref,
-) {
-  const [selfFocused, setSelfFocused] = useState(false);
-  const fallbackId = useId();
-  const inputId = id ?? fallbackId;
-
-  // `focused` is an optional controlled override, matching the original API.
-  const isFocused = focused !== undefined ? focused : selfFocused;
-
-  const hasValue = typeof value === "string" && value.length > 0;
-  const showsClearButton = showClearButton && hasValue && !disabled;
-  const showsInstructions = instructions
-    ? !(errors && hideInstructionsOnErrors)
-    : false;
-
-  const handleFocus = useCallback(
-    (event) => {
-      setSelfFocused(true);
-      if (onFocus) onFocus(event);
-    },
-    [onFocus],
-  );
-
-  const handleBlur = useCallback(
-    (event) => {
-      setSelfFocused(false);
-      if (onBlur) onBlur(event);
-    },
-    [onBlur],
-  );
-
-  const instructionsId = `${inputId}-instructions`;
-  const errorsId = `${inputId}-errors`;
-  const describedBy =
-    clsx(
-      ariaDescribedBy,
-      showsInstructions && instructionsId,
-      errors && errorsId,
-    ) || undefined;
-
-  return (
-    <div
-      className={clsx("hig__text-field-v1", {
-        "hig__text-field-v1--button": type === "button",
-        "hig__text-field-v1--required": required,
-        "hig__text-field-v1--disabled": disabled,
-        "hig__text-field-v1--clear-button-visible": showsClearButton,
-        "hig__text-field-v1--with-errors": errors,
-      })}
-    >
-      <div
-        className={clsx("hig__text-field-v1__content", {
-          "hig__text-field-v1__content--with-icon": icon,
-        })}
-      >
-        <div
-          className={clsx("hig__text-field-v1__input-wrapper", {
-            "hig__text-field-v1__input-wrapper--focused": isFocused,
-            "hig__text-field-v1__input-wrapper--with-errors": errors,
-            "hig__text-field-v1__input-wrapper--disabled": disabled,
-          })}
-        >
-          {label && <span className="hig__text-field-v1__label-spacer" />}
-          {label && (
-            <label
-              htmlFor={inputId}
-              className={clsx("hig__text-field-v1__label", {
-                "hig__text-field-v1__label--input-focused": isFocused,
-                "hig__text-field-v1__label--required": required,
-                "hig__text-field-v1__label--with-value": hasValue,
-              })}
-            >
-              {label}
-            </label>
-          )}
-          <div className="hig__text-field-v1__input-row">
-            {icon && (
-              <span
-                className={clsx("hig__text-field-v1__icon", {
-                  "hig__text-field-v1__icon--disabled": disabled,
-                })}
-              >
-                {icon}
-              </span>
-            )}
-            <input
-              {...inputProps}
-              aria-describedby={describedBy}
-              aria-invalid={errors ? true : undefined}
-              className={clsx(
-                "hig__text-field-v1__input",
-                inputProps.className,
-                { "hig__text-field-v1__input--no-value": !hasValue },
-              )}
-              disabled={disabled}
-              id={inputId}
-              onBlur={handleBlur}
-              onFocus={handleFocus}
-              readOnly
-              ref={ref}
-              required={required ? true : undefined}
-              type={type}
-              value={value ?? ""}
-            />
-            {showsClearButton && (
-              <span className="hig__text-field-v1__clear">
-                <button
-                  className="hig__icon-button hig__icon-button--transparent"
-                  onClick={onClearButtonClick}
-                  title={clearButtonTitle}
-                  type="button"
-                >
-                  <span className="hig__icon-button__icon">
-                    <ClearIcon />
-                  </span>
-                  <span className="hig__sr-only">{clearButtonTitle}</span>
-                </button>
-              </span>
-            )}
-          </div>
-        </div>
-        {showsInstructions && (
-          <p className="hig__text-field-v1__instructions" id={instructionsId}>
-            {instructions}
-          </p>
-        )}
-        {errors && (
-          <p className="hig__text-field-v1__errors" id={errorsId}>
-            {errors}
-          </p>
-        )}
-        {required && (
-          <p className="hig__text-field-v1__required-notice">{required}</p>
-        )}
-      </div>
-    </div>
-  );
-});
-
-export default TextField;
