--- src/index.js
@@ -1,1 +1,0 @@
-export { default, default as DatePicker } from "./DatePicker.jsx";
