+++ src/index.ts
@@ -1,0 +1,7 @@
+export { default, default as DatePicker } from "./DatePicker.js";
+export type {
+  DateLike,
+  DateLikeWithMessage,
+  DatePickerProps,
+} from "./DatePicker.js";
+export type { TextFieldProps } from "./TextField.js";
