+++ tsconfig.json
@@ -1,16 +1,36 @@
 {
   "compilerOptions": {
     "target": "ES2020",
-    "lib": ["ES2020", "DOM", "DOM.Iterable"],
+    "lib": [
+      "ES2020",
+      "DOM",
+      "DOM.Iterable"
+    ],
     "module": "ESNext",
     "moduleResolution": "bundler",
     "jsx": "react-jsx",
+    "types": [
+      "vite/client"
+    ],
     "strict": true,
-    "skipLibCheck": true,
-    "noEmit": true,
-    "allowJs": false,
+    "noUncheckedIndexedAccess": true,
+    "noImplicitOverride": true,
+    "noFallthroughCasesInSwitch": true,
+    "exactOptionalPropertyTypes": false,
+    "isolatedModules": true,
+    "verbatimModuleSyntax": true,
     "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true
+    "resolveJsonModule": true,
+    "forceConsistentCasingInFileNames": true,
+    "skipLibCheck": true,
+    "noEmit": true
   },
-  "include": ["src/index.d.ts", "types-test.ts"]
+  "include": [
+    "src",
+    "e2e",
+    "demo/src",
+    "*.config.ts",
+    "vitest.setup.ts",
+    "types-test.ts"
+  ]
 }
