+++ vite.demo.config.ts
