--- vitest.config.js
@@ -1,17 +1,0 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
-
-export default defineConfig({
-  plugins: [react()],
-  test: {
-    environment: "jsdom",
-    globals: true,
-    setupFiles: ["./vitest.setup.js"],
-    include: ["src/**/*.test.{js,jsx}"],
-    coverage: {
-      provider: "v8",
-      include: ["src/**/*.{js,jsx}"],
-      exclude: ["src/**/*.test.{js,jsx}"],
-    },
-  },
-});
