--- vitest.setup.js
@@ -1,5 +1,0 @@
-import "@testing-library/jest-dom/vitest";
-import { expect } from "vitest";
-import * as matchers from "vitest-axe/matchers";
-
-expect.extend(matchers);
